Add brute-force truncate-by-copy to RdBackFile.Close.
The Modula-3 libraries provide no way to truncate a RegularFile,
which is needed here.  Adding this would entail OS-dependent
versions for POSIX and WIN32 and make this module dependent on having
the latest RegularFile in the CM3 installation.  I don't what to
get into that now, so am resorting to doing a partial copy of the
disk file.  Yes, it's crude, inefficient, and entails a lot of
code to acheive such a simple result, but it's the path of least
resistance now.

Plus odds & ends.