Accidental damage repair: libvararray->vararray->libvararray
This happened as result of rsync to hudson, edits to other files,
commit and push, back on ignitron, pull.  I guess I need to
completely delete before rsync to a different computer.

Thare are two shipped cm3 packages: vararray and libvararray.
The latter is the one in use in FM3.  The differences are minimal,
need to be investigated, and one or the other removed, to prevent
more trouble.