Auto-create build dir & rework logs.
It started out creating build directory, but testing dragged me into
unravelling confusion about the two kinds of logs, where they
are created, etc.

There seems no way in libm3 to do a create-dir-if-does-not-already-exist,
nor to pretest directory existence.

This uses a Mickey-Mouse technique of create, catch exception, and
check the exception arg for an already-existed error, in a way that
is undoubtedly OS-dependent.  There could be other reasons for create
to fail, that should evoke messages.