WIP Rework scope entry/exit in anticipation of separate compilation.
Not complete, not compiled.

Use two separate stacks of scopes, one for where idents are being
declared (in the top scope only) and one for how they are looked up
(possibly anywhere in the whole stack).  These are different during
handling of exports and imports. This will also be needed during
record and object type definitions and procedure signatures.