OrdSets & file prefix.
Consistify on OrdSets that uses ElemT instead of ValidElemT as formal
to callbacks.  lalr uses this too, but where it is getting OrdSets from,
I didn't track down.

Rework handling of FM3 file prefixes and centralize in
sharedsrc.  But RdBackFile will either have to duplicate it or
bring in sharedsrc, which it does not now do.  That could open up
pandora's box.  Also RdBackFile needs to read and write as a single
random access read/write file, instead of using Wr.T and Rd.T.  So
it may need to both duplicate and bring in.

Also change the prefix to have a version character, cutting the magic
number down to 3 bytes.  RdBackFile appears to already be doing this.